COPR/requirements.txt
absl-py==0.11.0
albumentations==0.5.2
antlr4-python3-runtime==4.8
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
blessings==1.7
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
ConfigArgParse==1.2.3
cupy-cuda102==8.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastrlock==0.5
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
gpustat==0.6.0
grpcio==1.33.2
hydra-core==1.0.4
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==2.0.0
importlib-resources==3.3.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.2.0
kornia==0.4.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mkl-fft==1.2.0
mkl-random #1.1.1
mkl-service==2.3.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy #@ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233707986/work
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
omegaconf==2.0.5
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
opencv-python-headless==4.4.0.46
packaging==20.8
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow #@ file:///tmp/build/80754af9/pillow_1594307325547/work
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
pynvrtc==9.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-metric-learning==0.9.94
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.50.2
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.0
faiss #==1.6.3
